created admin user profile

+++ app/Http/Controllers/CategoriesController.php
@@ -96,6 +96,10 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->post->count() > 0) {
+            session()->flash('error', 'Category cannot be deleted because it has some posts.');
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully.');
